command pattern

+++ cart/command.py
@@ -1,25 +1,57 @@
 from dataclasses import dataclass
 from .models import Cart,CartItem,Order,OrderItem
+from django.contrib.auth.models import User
 from typing import Union
 from product.models import SpecialEditionGame,DLC,ProductDecorator
+from django.shortcuts import get_object_or_404
 
 
 @dataclass 
 class AddToOrderCommand:
     order: Order
-    item: Union[SpecialEditionGame,DLC,ProductDecorator]
+    items: Union[SpecialEditionGame,DLC,ProductDecorator]
     
-    def execute(self):
-        self.order.add_item(item=self.item)
+    def execute(self) -> None:
+        if self.items is None: 
+            return
+        self.order.add_item(items=self.items)
+    
+    def undo(self) -> None:
+        if self.items is None: 
+            return
+        self.order.delete_item(item=self.item)
         
 @dataclass
 class RemoveFromOrderCommand:
     order: Order
-    item: Union[SpecialEditionGame,DLC,ProductDecorator]
     
-    def execute(self):
-        self.order.delete_item(item=self.item)
+    def execute(self) -> None:
+        self.order.delete_item()
+    
+    def undo(self) -> None:
+        self.order.add_item(item=self.item)
 
+@dataclass
+class CreateOrderCommand:
+    user: User
+    transaction_id: float
+    def execute(self) -> Order:
+        order, created = Order.objects.get_or_create(user=self.user, transaction_id=self.transaction_id)
+        return order
+    def undo(self):
+        order = get_object_or_404(Order, user=self.user, transaction_id=self.transaction_id)
+        order.delete()
+
+@dataclass
+class DeleteOrderCommand:
+    user: User
+    transaction_id: float
+    def execute(self) -> None:
+        order = get_object_or_404(Order, user=self.user, transaction_id=self.transaction_id)
+        order.delete()
+    def undo(self) -> Order:
+        order, created = Order.objects.get_or_create(user=self.user, transaction_id=self.transaction_id)
+        return order
